Add readme to file folder
There was only one warning when adding the readme and doing
data-plane validation. There was a naming conflict between the Share_
prefix in some of the operationIds and the Share definition. Changed
Share and File operations to use the plural noun. Did not change
Directory because there is no naming conflict and I suspect that
complications may arise from the fact that the plural of Directory is
not just Directory with an 's' at the end.